ajuste entidades y referencia a bd de nml

+++ src/main/java/mx/gob/imss/cit/mjlssc/persistence/repository/MjltAsuntoActorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import mx.gob.imss.cit.mjlssc.persistence.entity.MjltAsuntoActor;
 
 @Repository
-public interface MjltAsuntoActorRepository extends JpaRepository<MjltAsuntoActor, Integer> {
+public interface MjltAsuntoActorRepository extends JpaRepository<MjltAsuntoActor, Long> {
 
 }
